Refactor EventHandler for less code lines
